+++ src/main/java/com/purbon/kafka/topology/KSqlArtefactManager.java
@@ -33,162 +33,135 @@ import org.apache.logging.log4j.Logger;
 
 public class KSqlArtefactManager extends ArtefactManager {
 
-  private static final Logger LOGGER = LogManager.getLogger(KSqlArtefactManager.class);
+    private static final Logger LOGGER = LogManager.getLogger(KSqlArtefactManager.class);
 
-  public KSqlArtefactManager(
-      ArtefactClient client, Configuration config, String topologyFileOrDir) {
-    super(client, config, topologyFileOrDir);
-  }
-
-  public KSqlArtefactManager(
-      Map<String, ? extends ArtefactClient> clients,
-      Configuration config,
-      String topologyFileOrDir) {
-    super(clients, config, topologyFileOrDir);
-  }
-
-  @Override
-  Collection<? extends Artefact> loadActualClusterStateIfAvailable(ExecutionPlan plan)
-      throws IOException {
-    var currentState =
-        config.fetchStateFromTheCluster() ? getClustersState() : plan.getKSqlArtefacts();
-
-    if (!config.fetchStateFromTheCluster()) {
-      // should detect if there are divergences between the local cluster state and the current
-      // status in the cluster
-      detectDivergencesInTheRemoteCluster(plan);
+    public KSqlArtefactManager(
+            ArtefactClient client, Configuration config, String topologyFileOrDir) {
+        super(client, config, topologyFileOrDir);
     }
 
-    return currentState;
-  }
-
-  private void detectDivergencesInTheRemoteCluster(ExecutionPlan plan) throws IOException {
-    var remoteArtefacts = getClustersState();
-
-    var delta =
-        plan.getKSqlArtefacts().stream()
-            .filter(localArtefacts -> !remoteArtefacts.contains(localArtefacts))
-            .collect(Collectors.toList());
-
-    if (delta.size() > 0) {
-      String errorMessage =
-          "Your remote state has changed since the last execution, these ksqlDB Artifact(s): "
-              + StringUtils.join(delta, ",")
-              + " are in your local state, but not in the cluster, please investigate!";
-      LOGGER.error(errorMessage);
-      throw new IOException(errorMessage);
-    }
-  }
-
-  @Override
-  protected List<? extends Artefact> findArtefactsToBeDeleted(
-      Collection<? extends Artefact> currentArtefacts, Set<Artefact> artefacts) {
-
-    var artefactsList =
-        currentArtefacts.stream()
-            .filter(a -> !artefacts.contains(a))
-            .sorted((o1, o2) -> -1 * ((KsqlArtefact) o1).compareTo((KsqlArtefact) o2))
-            .collect(Collectors.toCollection(LinkedList::new));
-
-    Map<String, LinkedList<KsqlArtefact>> artefactsMap = new HashMap<>();
-    artefactsMap.put("table", new LinkedList<>());
-    artefactsMap.put("stream", new LinkedList<>());
-
-    artefactsList.forEach(
-        (Consumer<Artefact>)
-            artefact -> {
-              if (artefact instanceof KsqlTableArtefact) {
-                artefactsMap.get("table").add((KsqlArtefact) artefact);
-              } else {
-                artefactsMap.get("stream").add((KsqlArtefact) artefact);
-              }
-            });
-
-    LinkedList<KsqlArtefact> toDeleteArtefactsList = new LinkedList<>();
-    for (String key : Arrays.asList("table", "stream")) {
-      artefactsMap.get(key).descendingIterator().forEachRemaining(toDeleteArtefactsList::add);
-    }
-    return toDeleteArtefactsList;
-  }
-
-  private Collection<? extends Artefact> getClustersState() throws IOException {
-    List<Either> list =
-        clients.values().stream()
-            .map(
-                client -> {
-                  try {
-                    Collection<? extends Artefact> artefacts = client.getClusterState();
-                    if (artefacts.isEmpty()) {
-                      return Either.Right(null);
-                    }
-                    return Either.Right(artefacts);
-                  } catch (IOException ex) {
-                    return Either.Left(ex);
-                  }
-                })
-            .collect(Collectors.toList());
-
-    List<IOException> errors =
-        list.stream()
-            .filter(Either::isLeft)
-            .map(e -> (IOException) e.getLeft().get())
-            .collect(Collectors.toList());
-    if (errors.size() > 0) {
-      throw new IOException(errors.get(0));
+    public KSqlArtefactManager(
+            Map<String, ? extends ArtefactClient> clients,
+            Configuration config,
+            String topologyFileOrDir) {
+        super(clients, config, topologyFileOrDir);
     }
 
-    return list.stream()
-        .filter(Either::isRight)
-        .flatMap(
-            (Function<Either, Stream<? extends Artefact>>)
-                either -> {
-                  Collection<? extends Artefact> artefacts =
-                      (Collection<? extends Artefact>) either.getRight().get();
-                  return artefacts.stream();
-                })
-        .map(
-            artefact -> {
-              if (artefact instanceof KsqlStreamArtefact) {
-                return new KsqlStreamArtefact(artefact.getPath(), null, artefact.getName());
-              } else if (artefact instanceof KsqlTableArtefact) {
-                return new KsqlTableArtefact(artefact.getPath(), null, artefact.getName());
-              } else {
-                LOGGER.error("KSQL Artefact of wrong type " + artefact.getClass());
-                return null;
-              }
-            })
-        .filter(Objects::nonNull)
-        .collect(Collectors.toSet());
-  }
+    @Override
+    protected Collection<? extends Artefact> getLocalState(ExecutionPlan plan) {
+        return plan.getKSqlArtefacts();
+    }
 
-  @Override
-  Set<Artefact> parseNewArtefacts(Topology topology) {
-    return topology.getProjects().stream()
-        .flatMap(
-            (Function<Project, Stream<Artefact>>)
-                project -> {
-                  KsqlArtefacts kSql = project.getKsqlArtefacts();
-                  return Stream.concat(kSql.getStreams().stream(), kSql.getTables().stream());
-                })
-        .sorted()
-        .collect(Collectors.toCollection(LinkedHashSet::new));
-  }
+    @Override
+    protected List<? extends Artefact> findArtefactsToBeDeleted(
+            Collection<? extends Artefact> currentArtefacts, Set<Artefact> artefacts) {
 
-  @Override
-  boolean isAllowDelete() {
-    return config.isAllowDeleteKsqlArtefacts();
-  }
+        var artefactsList =
+                currentArtefacts.stream()
+                        .filter(a -> !artefacts.contains(a))
+                        .sorted((o1, o2) -> -1 * ((KsqlArtefact) o1).compareTo((KsqlArtefact) o2))
+                        .collect(Collectors.toCollection(LinkedList::new));
 
-  @Override
-  String rootPath() {
-    return Files.isDirectory(Paths.get(topologyFileOrDir))
-        ? topologyFileOrDir
-        : new File(topologyFileOrDir).getParent();
-  }
+        Map<String, LinkedList<KsqlArtefact>> artefactsMap = new HashMap<>();
+        artefactsMap.put("table", new LinkedList<>());
+        artefactsMap.put("stream", new LinkedList<>());
 
-  @Override
-  public void printCurrentState(PrintStream out) throws IOException {
-    out.println("List of KSQL Artifacts:");
-    getClustersState().forEach(out::println);
-  }
+        artefactsList.forEach(
+                (Consumer<Artefact>)
+                        artefact -> {
+                            if (artefact instanceof KsqlTableArtefact) {
+                                artefactsMap.get("table").add((KsqlArtefact) artefact);
+                            } else {
+                                artefactsMap.get("stream").add((KsqlArtefact) artefact);
+                            }
+                        });
+
+        LinkedList<KsqlArtefact> toDeleteArtefactsList = new LinkedList<>();
+        for (String key : Arrays.asList("table", "stream")) {
+            artefactsMap.get(key).descendingIterator().forEachRemaining(toDeleteArtefactsList::add);
+        }
+        return toDeleteArtefactsList;
+    }
+
+    @Override
+    protected Collection<? extends Artefact> getClustersState() throws IOException {
+        List<Either> list =
+                clients.values().stream()
+                        .map(
+                                client -> {
+                                    try {
+                                        Collection<? extends Artefact> artefacts = client.getClusterState();
+                                        if (artefacts.isEmpty()) {
+                                            return Either.Right(null);
+                                        }
+                                        return Either.Right(artefacts);
+                                    } catch (IOException ex) {
+                                        return Either.Left(ex);
+                                    }
+                                })
+                        .collect(Collectors.toList());
+
+        List<IOException> errors =
+                list.stream()
+                        .filter(Either::isLeft)
+                        .map(e -> (IOException) e.getLeft().get())
+                        .collect(Collectors.toList());
+        if (errors.size() > 0) {
+            throw new IOException(errors.get(0));
+        }
+
+        return list.stream()
+                .filter(Either::isRight)
+                .flatMap(
+                        (Function<Either, Stream<? extends Artefact>>)
+                                either -> {
+                                    Collection<? extends Artefact> artefacts =
+                                            (Collection<? extends Artefact>) either.getRight().get();
+                                    return artefacts.stream();
+                                })
+                .map(
+                        artefact -> {
+                            if (artefact instanceof KsqlStreamArtefact) {
+                                return new KsqlStreamArtefact(artefact.getPath(), null, artefact.getName());
+                            } else if (artefact instanceof KsqlTableArtefact) {
+                                return new KsqlTableArtefact(artefact.getPath(), null, artefact.getName());
+                            } else {
+                                LOGGER.error("KSQL Artefact of wrong type " + artefact.getClass());
+                                return null;
+                            }
+                        })
+                .filter(Objects::nonNull)
+                .collect(Collectors.toSet());
+    }
+
+    @Override
+    Set<Artefact> parseNewArtefacts(Topology topology) {
+        return topology.getProjects().stream()
+                .flatMap(
+                        (Function<Project, Stream<Artefact>>)
+                                project -> {
+                                    KsqlArtefacts kSql = project.getKsqlArtefacts();
+                                    return Stream.concat(kSql.getStreams().stream(), kSql.getTables().stream());
+                                })
+                .sorted()
+                .collect(Collectors.toCollection(LinkedHashSet::new));
+    }
+
+    @Override
+    boolean isAllowDelete() {
+        return config.isAllowDeleteKsqlArtefacts();
+    }
+
+    @Override
+    String rootPath() {
+        return Files.isDirectory(Paths.get(topologyFileOrDir))
+                ? topologyFileOrDir
+                : new File(topologyFileOrDir).getParent();
+    }
+
+    @Override
+    public void printCurrentState(PrintStream out) throws IOException {
+        out.println("List of KSQL Artifacts:");
+        getClustersState().forEach(out::println);
+    }
 }
