+++ src/main/java/com/purbon/kafka/topology/KafkaConnectArtefactManager.java
@@ -28,45 +28,18 @@ public class KafkaConnectArtefactManager extends ArtefactManager {
     super(client, config, topologyFileOrDir);
   }
 
+  @Override
+  protected Collection<? extends Artefact> getLocalState(ExecutionPlan plan) {
+    return plan.getConnectors();
+  }
+
   public KafkaConnectArtefactManager(
       Map<String, KConnectApiClient> clients, Configuration config, String topologyFileOrDir) {
     super(clients, config, topologyFileOrDir);
   }
 
   @Override
-  Collection<? extends Artefact> loadActualClusterStateIfAvailable(ExecutionPlan plan)
-      throws IOException {
-    var currentState =
-        config.fetchStateFromTheCluster() ? getClustersState() : plan.getConnectors();
-
-    if (!config.fetchStateFromTheCluster()) {
-      // should detect if there are divergences between the local cluster state and the current
-      // status in the cluster
-      detectDivergencesInTheRemoteCluster(plan);
-    }
-
-    return currentState;
-  }
-
-  private void detectDivergencesInTheRemoteCluster(ExecutionPlan plan) throws IOException {
-    var remoteConnectors = getClustersState();
-
-    var delta =
-        plan.getConnectors().stream()
-            .filter(localConnector -> !remoteConnectors.contains(localConnector))
-            .collect(Collectors.toList());
-
-    if (delta.size() > 0) {
-      String errorMessage =
-          "Your remote state has changed since the last execution, these Connector(s): "
-              + StringUtils.join(delta, ",")
-              + " are in your local state, but not in the cluster, please investigate!";
-      LOGGER.error(errorMessage);
-      throw new IOException(errorMessage);
-    }
-  }
-
-  private Collection<? extends Artefact> getClustersState() throws IOException {
+  protected Collection<? extends Artefact> getClustersState() throws IOException {
     List<Either> list =
         clients.values().stream()
             .map(
